Refactors Ui into packages
