+++ src/main/java/trackr/ui/MainWindow.java
@@ -17,6 +17,7 @@ import trackr.logic.Logic;
 import trackr.logic.commands.CommandResult;
 import trackr.logic.commands.exceptions.CommandException;
 import trackr.logic.parser.exceptions.ParseException;
+import trackr.ui.dashboard.TabPanel;
 
 /**
  * The Main Window. Provides the basic application layout containing
