+++ src/main/java/trackr/ui/cards/TaskCard.java
@@ -1,10 +1,11 @@
-package trackr.ui;
+package trackr.ui.cards;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import trackr.model.task.Task;
+import trackr.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Task}.
