+++ src/main/java/trackr/ui/dashboard/HomeView.java
@@ -1,4 +1,4 @@
-package trackr.ui;
+package trackr.ui.dashboard;
 
 import javafx.fxml.FXML;
 import javafx.scene.layout.GridPane;
@@ -8,6 +8,9 @@ import javafx.scene.layout.Region;
 import javafx.scene.layout.StackPane;
 import javafx.scene.text.Text;
 import trackr.logic.Logic;
+import trackr.ui.UiPart;
+import trackr.ui.listpanels.SupplierListPanel;
+import trackr.ui.listpanels.TaskListPanel;
 
 /**
  * The UI component that is resposnible for displaying Home tab.
