+++ src/main/java/trackr/ui/dashboard/TabPanel.java
@@ -1,4 +1,4 @@
-package trackr.ui;
+package trackr.ui.dashboard;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Tab;
@@ -6,8 +6,13 @@ import javafx.scene.control.TabPane;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.StackPane;
 import javafx.scene.text.Text;
+
 import trackr.commons.core.index.Index;
 import trackr.logic.Logic;
+import trackr.ui.UiPart;
+import trackr.ui.listpanels.OrderListPanel;
+import trackr.ui.listpanels.SupplierListPanel;
+import trackr.ui.listpanels.TaskListPanel;
 
 /**
  * The UI component that is responsible for displaying Tab views.
